Move pointing bind from HandsSystem to PointingSystem
